requirements.txt
# prev deps…
pandas>=2.0
numpy>=1.24
scikit-learn>=1.3
PyYAML>=6.0
joblib>=1.3
vaderSentiment>=3.3.2
optuna>=3.4.0
matplotlib>=3.8
# LLM deps
torch>=2.1,<3.0
transformers>=4.44
accelerate>=0.33
